requirements.txt
streamlit
google-cloud-bigquery
pandas
langchain-openai
langchain-core
langchain-community
